+++ examples/html-py-ever/tests/conftest.py
@@ -1,4 +1,5 @@
 import sys
+
 import pytest
 
 if sys.platform == "emscripten":