+++ examples/rust_with_cffi/cffi_module.py
@@ -1,6 +1,5 @@
 import cffi
 
-
 ffi = cffi.FFI()
 ffi.cdef(
     """