+++ setuptools_rust/rustc_info.py
@@ -2,9 +2,10 @@ from __future__ import annotations
 
 import os
 import subprocess
-from setuptools.errors import PlatformError
 from functools import lru_cache
-from typing import Dict, List, NewType, Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING, NewType
+
+from setuptools.errors import PlatformError
 
 from ._utils import Env, check_subprocess_output
 
@@ -12,7 +13,7 @@ if TYPE_CHECKING:
     from semantic_version import Version
 
 
-def get_rust_version(env: Optional[Env]) -> Optional[Version]:  # type: ignore[no-any-unimported]
+def get_rust_version(env: Env | None) -> Version | None:  # type: ignore[no-any-unimported]
     try:
         # first line of rustc -Vv is something like
         # rustc 1.61.0 (fe5b13d68 2022-05-18)
@@ -26,7 +27,7 @@ def get_rust_version(env: Optional[Env]) -> Optional[Version]:  # type: ignore[n
 _HOST_LINE_START = "host: "
 
 
-def get_rust_host(env: Optional[Env]) -> str:
+def get_rust_host(env: Env | None) -> str:
     # rustc -Vv has a line denoting the host which cargo uses to decide the
     # default target, e.g.
     # host: aarch64-apple-darwin
@@ -36,7 +37,7 @@ def get_rust_host(env: Optional[Env]) -> str:
     raise PlatformError("Could not determine rust host")
 
 
-RustCfgs = NewType("RustCfgs", Dict[str, Optional[str]])
+RustCfgs = NewType("RustCfgs", dict[str, str | None])
 
 
 def _is_custom_target(target: str) -> bool:
@@ -52,7 +53,7 @@ def _is_custom_target(target: str) -> bool:
     return False
 
 
-def get_rustc_cfgs(target_triple: Optional[str], env: Env) -> RustCfgs:
+def get_rustc_cfgs(target_triple: str | None, env: Env) -> RustCfgs:
     cfgs = RustCfgs({})
     for entry in get_rust_target_info(target_triple, env):
         maybe_split = entry.split("=", maxsplit=1)
@@ -64,8 +65,8 @@ def get_rustc_cfgs(target_triple: Optional[str], env: Env) -> RustCfgs:
     return cfgs
 
 
-@lru_cache()
-def get_rust_target_info(target_triple: Optional[str], env: Env) -> List[str]:
+@lru_cache
+def get_rust_target_info(target_triple: str | None, env: Env) -> list[str]:
     cmd = ["rustc", "--print", "cfg"]
     if target_triple:
         if _is_custom_target(target_triple):
@@ -75,19 +76,19 @@ def get_rust_target_info(target_triple: Optional[str], env: Env) -> List[str]:
     return output.splitlines()
 
 
-@lru_cache()
-def get_rust_target_list(env: Env) -> List[str]:
+@lru_cache
+def get_rust_target_list(env: Env) -> list[str]:
     output = check_subprocess_output(
         ["rustc", "--print", "target-list"], env=env, text=True
     )
     return output.splitlines()
 
 
-@lru_cache()
+@lru_cache
 def _rust_version(env: Env) -> str:
     return check_subprocess_output(["rustc", "-V"], env=env, text=True)
 
 
-@lru_cache()
+@lru_cache
 def _rust_version_verbose(env: Env) -> str:
     return check_subprocess_output(["rustc", "-Vv"], env=env, text=True)